collectors/bluesky/requirements.txt
functions-framework==3.5.0
requests==2.31.0
psycopg2-binary==2.9.9
